requirements.txt
streamlit
pytesseract 
pillow
python-dotenv
langchain
PyPDF2
chromadb
google-generativeai
googletrans==4.0.0-rc1